requirements.txt
# Core project currently chỉ dùng thư viện chuẩn Python.
# Các gói bên dưới phục vụ cho thư mục Mina_Core (Google Drive + Gemini).

google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
google-generativeai>=0.5.0

# Recommended Python version:
python>=3.10
